Removed yield from _beat, reducing lag on Linux.
Mixing Desk used yield in _beat to prevent processing a beat for multiple
times. The approach introduces a lag that is quite noticeable on Linux.

This commit removed the yield call, and tracked triggered beats in a variable
instead. This allows beats to be tracked more precisely, reducing the lag.